Check to make sure children are not nested lists.
